# docs/swagger.yaml
basePath: /api/v1/
definitions:
  controller.CartResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/db.FindCartRow'
        type: array
      total_quantity:
        type: integer
    type: object
  controller.CreateCartRequest:
    properties:
      product_id:
        type: string
    type: object
  controller.CreateCategoryRequest:
    properties:
      category_name:
        type: string
    type: object
  controller.CreateProductRequest:
    properties:
      category_id:
        type: integer
      price:
        type: integer
      product_name:
        type: string
    type: object
  controller.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  controller.RegisterRequest:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  controller.Response:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  db.Category:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
    type: object
  db.FindCartRow:
    properties:
      cart_id:
        type: string
      customer_name:
        type: string
      price:
        type: integer
      product_name:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: An API Documentation Online Store App
  title: Synapsis Online Store API
  version: "1.0"
paths:
  /api/v1/products:
    get:
      consumes:
      - application/json
      description: Listing all of products by category id
      parameters:
      - description: category
        in: query
        name: category
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response'
      summary: Get Product by category id
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Adding new product data
      parameters:
      - description: add product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/controller.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response'
      summary: Add new product data
      tags:
      - Product
  /carts:
    get:
      consumes:
      - application/json
      description: Listing all of products that customer have been add.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CartResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - Authorization: []
      summary: Find existing cart.
      tags:
      - Shopping Cart
    post:
      consumes:
      - application/json
      description: Add Cart using JSON
      parameters:
      - description: add cart
        in: body
        name: cart
        required: true
        schema:
          $ref: '#/definitions/controller.CreateCartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.CartResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response'
      summary: Add product item into cart
      tags:
      - Shopping Cart
  /carts/{cart_id}:
    delete:
      description: Delete selected cart item by id.
      parameters:
      - description: Cart ID
        in: path
        name: cart_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CartResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - Authorization: []
      summary: Delete existing cart.
      tags:
      - Shopping Cart
  /categories:
    get:
      description: Listing all of categories product.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Category'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response'
      summary: Get product categories
      tags:
      - Product Category
    post:
      consumes:
      - application/json
      description: Adding category for product
      parameters:
      - description: add category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/controller.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.Category'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response'
      summary: Add category product data
      tags:
      - Product Category
  /login:
    post:
      consumes:
      - application/json
      description: Authorize customer and returning token.
      parameters:
      - description: check customer
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/controller.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response'
      summary: Check and authorize customer.
      tags:
      - Customer
  /register:
    post:
      consumes:
      - application/json
      description: Register new customer
      parameters:
      - description: add customer
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response'
      summary: Add new customer data
      tags:
      - Customer
swagger: "2.0"